+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Add "Outgoing trust" category to External contacts
+- Show user counts by team on the statistics page
 
 ### Changed